+++ lib/networking/network_manager.dart
@@ -15,38 +15,41 @@
 
 import 'dart:developer';
 import 'dart:io';
+import 'dart:math' as math;
 
 import 'package:dio/dio.dart';
-import 'package:wp_json_api/enums/wp_auth_type.dart';
-import 'package:wp_json_api/exceptions/empty_username_exception.dart';
-import 'package:wp_json_api/exceptions/existing_user_email_exception.dart';
-import 'package:wp_json_api/exceptions/existing_user_login_exception.dart';
-import 'package:wp_json_api/exceptions/incorrect_password_exception.dart';
-import 'package:wp_json_api/exceptions/invalid_email_exception.dart';
-import 'package:wp_json_api/exceptions/invalid_nonce_exception.dart';
-import 'package:wp_json_api/exceptions/invalid_params_exception.dart';
-import 'package:wp_json_api/exceptions/invalid_user_token_exception.dart';
-import 'package:wp_json_api/exceptions/invalid_username_exception.dart';
-import 'package:wp_json_api/exceptions/user_already_exist_exception.dart';
-import 'package:wp_json_api/exceptions/username_taken_exception.dart';
-import 'package:wp_json_api/exceptions/woocommerce_not_found_exception.dart';
-import 'package:wp_json_api/models/responses/wc_customer_info_response.dart';
-import 'package:wp_json_api/models/responses/wc_customer_updated_response.dart';
-import 'package:wp_json_api/models/responses/wc_points_and_rewards_calculate_points.dart';
-import 'package:wp_json_api/models/responses/wc_points_and_rewards_user.dart';
-import 'package:wp_json_api/models/responses/wp_nonce_response.dart';
-import 'package:wp_json_api/models/responses/wp_nonce_verified_response.dart';
-import 'package:wp_json_api/models/responses/wp_user_add_role_response.dart';
-import 'package:wp_json_api/models/responses/wp_user_delete_response.dart';
-import 'package:wp_json_api/models/responses/wp_user_info_updated_response.dart';
-import 'package:wp_json_api/models/responses/wp_user_login_response.dart';
-import 'package:wp_json_api/models/responses/wp_user_register_response.dart';
-import 'package:wp_json_api/models/responses/wp_user_remove_role_response.dart';
-import 'package:wp_json_api/models/responses/wp_user_reset_password_response.dart';
-import 'package:wp_json_api/models/wp_meta_meta.dart';
-import 'package:wp_json_api/wp_json_api.dart';
-import 'package:wp_json_api/enums/wp_route_type.dart';
-import 'package:wp_json_api/models/responses/wp_user_info_response.dart';
+import 'package:nylo_support/helpers/helper.dart';
+import '/enums/wp_auth_type.dart';
+import '/exceptions/empty_username_exception.dart';
+import '/exceptions/existing_user_email_exception.dart';
+import '/exceptions/existing_user_login_exception.dart';
+import '/exceptions/incorrect_password_exception.dart';
+import '/exceptions/invalid_email_exception.dart';
+import '/exceptions/invalid_nonce_exception.dart';
+import '/exceptions/invalid_params_exception.dart';
+import '/exceptions/invalid_user_token_exception.dart';
+import '/exceptions/invalid_username_exception.dart';
+import '/exceptions/user_already_exist_exception.dart';
+import '/exceptions/username_taken_exception.dart';
+import '/exceptions/woocommerce_not_found_exception.dart';
+import '/models/responses/wc_customer_info_response.dart';
+import '/models/responses/wc_customer_updated_response.dart';
+import '/models/responses/wc_points_and_rewards_calculate_points.dart';
+import '/models/responses/wc_points_and_rewards_user.dart';
+import '/models/responses/wp_nonce_response.dart';
+import '/models/responses/wp_nonce_verified_response.dart';
+import '/models/responses/wp_user_add_role_response.dart';
+import '/models/responses/wp_user_delete_response.dart';
+import '/models/responses/wp_user_info_updated_response.dart';
+import '/models/responses/wp_user_login_response.dart';
+import '/models/responses/wp_user_register_response.dart';
+import '/models/responses/wp_user_remove_role_response.dart';
+import '/models/responses/wp_user_reset_password_response.dart';
+import '/models/wp_meta_meta.dart';
+import '/models/wp_user.dart';
+import '/wp_json_api.dart';
+import '/enums/wp_route_type.dart';
+import '/models/responses/wp_user_info_response.dart';
 
 /// A networking class to manage all the APIs from "wp_json_api"
 class WPAppNetworkManager {
@@ -72,7 +75,8 @@ class WPAppNetworkManager {
       String? username,
       required String password,
       WPAuthType authType = WPAuthType.WpEmail,
-      String? tokenExpiryAt}) async {
+      String? tokenExpiryAt,
+      bool saveTokenToLocalStorage = true}) async {
     // Get nonce from WordPress
     WPNonceResponse wpNonceResponse = await wpNonce();
 
@@ -81,7 +85,7 @@ class WPAppNetworkManager {
     if (username != null) payload["username"] = username;
     if (email != null) payload["email"] = email;
     payload["password"] = password;
-    payload["nonce"] = wpNonceResponse.data!.nonce;
+    payload["nonce"] = wpNonceResponse.data?.nonce;
     if (tokenExpiryAt != null) payload["expiry"] = tokenExpiryAt;
 
     // send http request
@@ -93,9 +97,18 @@ class WPAppNetworkManager {
     );
 
     // return response
-    return _jsonHasBadStatus(json)
-        ? this._throwExceptionForStatusCode(json)
-        : WPUserLoginResponse.fromJson(json);
+    if (_jsonHasBadStatus(json)) {
+      return _throwExceptionForStatusCode(json);
+    }
+    WPUserLoginResponse wpUserLoginResponse =
+        WPUserLoginResponse.fromJson(json);
+    String? userToken = wpUserLoginResponse.data?.userToken;
+
+    if (userToken != null && saveTokenToLocalStorage) {
+      WpUser wpUser = WpUser.fromWPUserLoginResponse(wpUserLoginResponse);
+      await WPJsonAPI.wpLogin(wpUser);
+    }
+    return wpUserLoginResponse;
   }
 
   /// Sends a request to register a user in WordPress with the following
@@ -113,17 +126,25 @@ class WPAppNetworkManager {
   Future<WPUserRegisterResponse> wpRegister(
       {required String email,
       required String password,
-      required String username,
-      String? expiry}) async {
+      String? username,
+      String? expiry,
+      Map<String, dynamic>? args,
+      bool saveTokenToLocalStorage = true}) async {
     // Get nonce from WordPress
     WPNonceResponse wpNonceResponse = await wpNonce();
+
+    if (username == null) {
+      username = (email.replaceAll(RegExp(r'([@.])'), "")) + _randomString(4);
+    }
 
     // Creates payload for register
     Map<String, dynamic> payload = {
       "email": email,
       "password": password,
       "username": username,
-      "nonce": wpNonceResponse.data!.nonce
+      "wc_register": false,
+      "nonce": wpNonceResponse.data?.nonce,
+      "args": args
     };
     if (expiry != null) payload["expiry"] = expiry;
 
@@ -135,9 +156,77 @@ class WPAppNetworkManager {
     );
 
     // return response
-    return _jsonHasBadStatus(json)
-        ? this._throwExceptionForStatusCode(json)
-        : WPUserRegisterResponse.fromJson(json);
+    if (_jsonHasBadStatus(json)) {
+      return _throwExceptionForStatusCode(json);
+    }
+    WPUserRegisterResponse wPUserRegisterResponse =
+        WPUserRegisterResponse.fromJson(json);
+    String? userToken = wPUserRegisterResponse.data?.userToken;
+
+    if (userToken != null && saveTokenToLocalStorage) {
+      WpUser wpUser = WpUser.fromWPUserRegisterResponse(wPUserRegisterResponse);
+      await WPJsonAPI.wpLogin(wpUser);
+    }
+    return wPUserRegisterResponse;
+  }
+
+  /// Sends a request to register a user in WooCommerce with the following
+  /// parameters [username], [email] and [password].
+  /// You can optionally set an [expiry] for the token expiry like "+1 day".
+  ///
+  /// Returns a [WPUserRegisterResponse] future.
+  /// Throws an [UsernameTakenException] if username is taken
+  /// [InvalidNonceException] if nonce token is not valid
+  /// [ExistingUserLoginException] if user login exists
+  /// [ExistingUserEmailException] if that email is in use
+  /// [UserAlreadyExistException] if a user was found with the same email
+  /// [EmptyUsernameException] if the username field has empty
+  /// [Exception] if fails.
+  Future<WPUserRegisterResponse> wcRegister(
+      {required String email,
+      required String password,
+      String? username,
+      String? expiry,
+      Map<String, dynamic>? args,
+      bool saveTokenToLocalStorage = true}) async {
+    // Get nonce from WordPress
+    WPNonceResponse wpNonceResponse = await wpNonce();
+
+    if (username == null) {
+      username = (email.replaceAll(RegExp(r'([@.])'), "")) + _randomString(4);
+    }
+
+    // Creates payload for register
+    Map<String, dynamic> payload = {
+      "email": email,
+      "password": password,
+      "username": username,
+      "wc_register": true, // This is the key to register a user in WooCommerce
+      "nonce": wpNonceResponse.data?.nonce,
+      "args": args
+    };
+    if (expiry != null) payload["expiry"] = expiry;
+
+    // send http request
+    final json = await _http(
+      method: "POST",
+      url: _urlForRouteType(WPRouteType.UserRegister),
+      body: payload,
+    );
+
+    // return response
+    if (_jsonHasBadStatus(json)) {
+      return _throwExceptionForStatusCode(json);
+    }
+    WPUserRegisterResponse wPUserRegisterResponse =
+        WPUserRegisterResponse.fromJson(json);
+    String? userToken = wPUserRegisterResponse.data?.userToken;
+
+    if (userToken != null && saveTokenToLocalStorage) {
+      WpUser wpUser = WpUser.fromWPUserRegisterResponse(wPUserRegisterResponse);
+      await WPJsonAPI.wpLogin(wpUser);
+    }
+    return wPUserRegisterResponse;
   }
 
   /// Makes a request to get a valid nonce code
@@ -182,13 +271,13 @@ class WPAppNetworkManager {
   ///
   /// Returns a [WPUserInfoResponse] future.
   /// Throws an [Exception] if fails
-  Future<WPUserInfoResponse> wpGetUserInfo(String userToken) async {
+  Future<WPUserInfoResponse> wpGetUserInfo({String? userToken}) async {
     // send http request
     final json = await _http(
-      method: "POST",
-      url: _urlForRouteType(WPRouteType.UserInfo),
-      userToken: userToken,
-    );
+        method: "POST",
+        url: _urlForRouteType(WPRouteType.UserInfo),
+        userToken: userToken,
+        shouldAuthRequest: true);
 
     // return response
     return _jsonHasBadStatus(json)
@@ -202,8 +291,9 @@ class WPAppNetworkManager {
   ///
   /// Returns a [WPUserInfoUpdatedResponse] future.
   /// Throws an [Exception] if fails.
-  Future<WPUserInfoUpdatedResponse> wpUpdateUserInfo(userToken,
-      {String? firstName,
+  Future<WPUserInfoUpdatedResponse> wpUpdateUserInfo(
+      {String? userToken,
+      String? firstName,
       String? lastName,
       String? displayName,
       List<WpMetaData>? metaData}) async {
@@ -222,6 +312,7 @@ class WPAppNetworkManager {
       method: "POST",
       url: _urlForRouteType(WPRouteType.UserUpdateInfo),
       userToken: userToken,
+      shouldAuthRequest: true,
       body: payload,
     );
 
@@ -236,8 +327,8 @@ class WPAppNetworkManager {
   ///
   /// Returns a [WPUserInfoUpdatedResponse] future.
   /// Throws an [Exception] if fails.
-  Future<WPUserAddRoleResponse> wpUserAddRole(userToken,
-      {required String role}) async {
+  Future<WPUserAddRoleResponse> wpUserAddRole(
+      {required String role, String? userToken}) async {
     Map<String, dynamic> payload = {};
     payload["role"] = role;
 
@@ -246,6 +337,7 @@ class WPAppNetworkManager {
       method: "POST",
       url: _urlForRouteType(WPRouteType.UserAddRole),
       userToken: userToken,
+      shouldAuthRequest: true,
       body: payload,
     );
 
@@ -255,12 +347,18 @@ class WPAppNetworkManager {
         : WPUserAddRoleResponse.fromJson(json);
   }
 
+  /// Logs out the user by deleting the user token from the local storage
+  wpLogout() async {
+    await NyStorage.delete(WPJsonAPI.storageKey());
+  }
+
   /// Sends a request to delete a WordPress user. Include a valid
   /// [userToken] and an optional [reassign] argument to send a successful request.
   ///
   /// Returns a [WPUserDeleteResponse] future.
   /// Throws an [Exception] if fails.
-  Future<WPUserDeleteResponse> wpUserDelete(userToken, {int? reassign}) async {
+  Future<WPUserDeleteResponse> wpUserDelete(
+      {String? userToken, int? reassign}) async {
     Map<String, dynamic> payload = {};
     if (reassign != null) {
       payload["reassign"] = reassign;
@@ -271,6 +369,7 @@ class WPAppNetworkManager {
       method: "POST",
       url: _urlForRouteType(WPRouteType.UserDelete),
       userToken: userToken,
+      shouldAuthRequest: true,
       body: payload,
     );
 
@@ -285,8 +384,8 @@ class WPAppNetworkManager {
   ///
   /// Returns a [WPUserInfoUpdatedResponse] future.
   /// Throws an [Exception] if fails.
-  Future<WPUserRemoveRoleResponse> wpUserRemoveRole(userToken,
-      {required String role}) async {
+  Future<WPUserRemoveRoleResponse> wpUserRemoveRole(
+      {required String role, String? userToken}) async {
     Map<String, dynamic> payload = {};
     payload["role"] = role;
 
@@ -295,6 +394,7 @@ class WPAppNetworkManager {
       method: "POST",
       url: _urlForRouteType(WPRouteType.UserRemoveRole),
       userToken: userToken,
+      shouldAuthRequest: true,
       body: payload,
     );
 
@@ -308,19 +408,19 @@ class WPAppNetworkManager {
   ///
   /// Returns a [WCCustomerInfoResponse] future.
   /// Throws an [Exception] if fails.
-  Future<WPUserResetPasswordResponse> wpResetPassword(
-    String userToken, {
+  Future<WPUserResetPasswordResponse> wpResetPassword({
     required String password,
+    String? userToken,
   }) async {
     Map<String, dynamic> payload = {"password": password};
 
     // send http request
     final json = await _http(
-      method: "POST",
-      url: _urlForRouteType(WPRouteType.UserUpdatePassword),
-      body: payload,
-      userToken: userToken,
-    );
+        method: "POST",
+        url: _urlForRouteType(WPRouteType.UserUpdatePassword),
+        body: payload,
+        userToken: userToken,
+        shouldAuthRequest: true);
 
     // return response
     return _jsonHasBadStatus(json)
@@ -332,13 +432,13 @@ class WPAppNetworkManager {
   ///
   /// Returns a [WCCustomerInfoResponse] future.
   /// Throws an [Exception] if fails.
-  Future<WCCustomerInfoResponse> wcCustomerInfo(String userToken) async {
+  Future<WCCustomerInfoResponse> wcCustomerInfo({String? userToken}) async {
     // send http request
     final json = await _http(
-      method: "POST",
-      url: _urlForRouteType(WPRouteType.WCCustomerInfo),
-      userToken: userToken,
-    );
+        method: "POST",
+        url: _urlForRouteType(WPRouteType.WCCustomerInfo),
+        userToken: userToken,
+        shouldAuthRequest: true);
 
     // return response
     return _jsonHasBadStatus(json)
@@ -350,12 +450,14 @@ class WPAppNetworkManager {
   ///
   /// Returns a [WcPointsAndRewardUser] future.
   /// Throws an [Exception] if fails.
-  Future<WcPointsAndRewardUser> wcPointsAndRewardsUser(String userToken) async {
+  Future<WcPointsAndRewardUser> wcPointsAndRewardsUser(
+      {String? userToken}) async {
     // send http request
     final json = await _http(
       method: "POST",
       url: _urlForRouteType(WPRouteType.WCPointsAndRewardsUser),
       userToken: userToken,
+      shouldAuthRequest: true,
     );
 
     // return response
@@ -369,14 +471,15 @@ class WPAppNetworkManager {
   /// Returns a [WcPointsAndRewardCalculatePoints] future.
   /// Throws an [Exception] if fails.
   Future<WcPointsAndRewardCalculatePoints> wcPointsAndRewardsCalculatePoints(
-      String userToken,
-      {required double points}) async {
+      {required double points, String? userToken}) async {
     // send http request
     final json = await _http(
-        method: "POST",
-        url: _urlForRouteType(WPRouteType.WCPointsAndRewardsCalculatePoints),
-        userToken: userToken,
-        body: {"points": points});
+      method: "POST",
+      url: _urlForRouteType(WPRouteType.WCPointsAndRewardsCalculatePoints),
+      userToken: userToken,
+      shouldAuthRequest: true,
+      body: {"points": points},
+    );
 
     // return response
     return _jsonHasBadStatus(json)
@@ -389,8 +492,9 @@ class WPAppNetworkManager {
   ///
   /// Returns [WCCustomerUpdatedResponse] future.
   /// Throws an [Exception] if fails.
-  Future<WCCustomerUpdatedResponse> wcUpdateCustomerInfo(userToken,
-      {String? firstName,
+  Future<WCCustomerUpdatedResponse> wcUpdateCustomerInfo(
+      {String? userToken,
+      String? firstName,
       String? lastName,
       String? displayName,
       String? billingFirstName,
@@ -462,11 +566,11 @@ class WPAppNetworkManager {
 
     // send http request
     final json = await _http(
-      method: "POST",
-      url: _urlForRouteType(WPRouteType.WCCustomerUpdateInfo),
-      body: payload,
-      userToken: userToken,
-    );
+        method: "POST",
+        url: _urlForRouteType(WPRouteType.WCCustomerUpdateInfo),
+        body: payload,
+        userToken: userToken,
+        shouldAuthRequest: true);
 
     // return response
     return _jsonHasBadStatus(json)
@@ -483,7 +587,8 @@ class WPAppNetworkManager {
       {required String method,
       required String url,
       Map<String, dynamic>? body,
-      String? userToken}) async {
+      String? userToken,
+      bool shouldAuthRequest = false}) async {
     Response? response;
     if (method == "GET") {
       response = await dio.get(url);
@@ -493,6 +598,10 @@ class WPAppNetworkManager {
       };
       if (body == null) {
         body = {};
+      }
+      String? userTokenFromStorage = await WPJsonAPI.wpUserToken();
+      if (shouldAuthRequest && userTokenFromStorage != null) {
+        body.addAll({"token": userTokenFromStorage});
       }
       if (userToken != null) {
         body.addAll({"token": userToken});
@@ -632,6 +741,17 @@ class WPAppNetworkManager {
           return "";
         }
     }
+  }
+
+  /// Generates a random string of [length] characters.
+  String _randomString(int length) {
+    const chars = "abcdefghijklmnopqrstuvwxyz0123456789";
+    math.Random rnd = math.Random(DateTime.now().millisecondsSinceEpoch);
+    String result = "";
+    for (var i = 0; i < length; i++) {
+      result += chars[rnd.nextInt(chars.length)];
+    }
+    return result;
   }
 
   /// Throws an exception from the [json] status returned from payload.