+++ lib/wp_json_api.dart
@@ -15,15 +15,21 @@
 
 library wp_json_api;
 
-import 'package:wp_json_api/helpers/typedefs.dart';
-import 'package:wp_json_api/networking/network_manager.dart';
+import 'package:nylo_support/helpers/auth.dart';
+import 'package:nylo_support/helpers/helper.dart';
+import '/helpers/typedefs.dart';
+import '/models/wp_user.dart';
+import '/networking/network_manager.dart';
 
-/// The base class to initialize and use WPJSONAPI
+/// The version of the wp_json_api
+String _wpJsonAPIVersion = "3.5.0";
+
+/// The base class to initialize and use WPJsonAPI
 class WPJsonAPI {
-  /// Private constructor for WPJSONAPI
+  /// Private constructor for WPJsonAPI
   WPJsonAPI._privateConstructor();
 
-  /// Instance of WPJSONAPI
+  /// Instance of WPJsonAPI
   static final WPJsonAPI instance = WPJsonAPI._privateConstructor();
 
   /// The base url for the WordPress Site e.g. https://mysitewp.com
@@ -35,17 +41,68 @@ class WPJsonAPI {
   /// Default API root for your WordPress site
   String _apiPath = "/wp-json";
 
+  /// The version
+  static String get version => _wpJsonAPIVersion;
+
   /// Initialize and configure class interface.
   /// You can optional set [shouldDebug] == false to stop debugging
   /// [wpJsonPath] is the root path for accessing you sites WordPress APIs
   /// by default this should be "/wp-json".
-  initWith(
+  init(
       {required String baseUrl,
       String wpJsonPath = '/wp-json',
       bool shouldDebug = true}) {
     _setBaseApi(baseUrl: baseUrl);
     _setApiPath(path: wpJsonPath);
     _setShouldDebug(value: shouldDebug);
+  }
+
+  /// Login a user with the [WpUser]
+  static wpLogin(WpUser wpUser) async {
+    await Auth.set(wpUser, key: storageKey());
+  }
+
+  /// Logout a user
+  static wpLogout() async {
+    await Auth.remove(key: storageKey());
+  }
+
+  /// Authenticate a user if they are logged in
+  static wpAuth() async {
+    await Auth.loginModel(
+        WPJsonAPI.storageKey(), (data) => WpUser.fromJson(data));
+  }
+
+  /// Check if a user is logged in
+  static Future<bool> wpUserLoggedIn() async {
+    WpUser? _wpUser = await wpUser();
+    if (_wpUser == null) {
+      return false;
+    }
+    if (_wpUser.token == null) {
+      return false;
+    }
+    return true;
+  }
+
+  /// Returns the logged in user
+  static Future<WpUser?> wpUser() async {
+    return await NyStorage.read<WpUser>(storageKey(), modelDecoders: {
+      WpUser: (json) => WpUser.fromJson(json),
+    });
+  }
+
+  /// Returns the user ID of the logged in user
+  static Future<String?> wpUserId() async {
+    WpUser? _wpUser = await wpUser();
+    return _wpUser?.id.toString();
+  }
+
+  /// Get the token for the user
+  static Future<String?> wpUserToken() async {
+    WpUser? _wpUser = await wpUser();
+    if (_wpUser == null) return null;
+    return _wpUser.token;
   }
 
   /// Sets the base API in the class
@@ -78,4 +135,7 @@ class WPJsonAPI {
   api(RequestCallback request) async {
     return await request(WPAppNetworkManager.instance);
   }
+
+  /// Returns the storage key for the plugin
+  static String storageKey() => "wp_json_api";
 }